Some bugs, some improves...

+++ app/src/main/java/team/itis/ivm/data/Sound.java
@@ -4,10 +4,10 @@ public class Sound {
 
     public String path;
     public String id;
-    public int start;
-    public int end;
+    public float start;
+    public float end;
 
-    public Sound(String path, int start, int end) {
+    public Sound(String path, float start, float end) {
         this.path = path;
         this.start = start;
         this.end = end;
